feat(component): 新增设备列表公共组件
